requirements.txt
pygame
pyautogui
Pillow
PySide6
PyQt6-tools
